+++ crates/ruma/src/lib.rs
@@ -85,9 +85,9 @@ pub use ruma_identifiers::{
     device_id, device_key_id, event_id, mxc_uri, room_alias_id, room_id, room_version_id,
     server_name, server_signing_key_id, user_id, ClientSecret, DeviceId, DeviceKeyAlgorithm,
     DeviceKeyId, DeviceSignatures, DeviceSigningKeyId, EntitySignatures, EventEncryptionAlgorithm,
-    EventId, KeyId, KeyName, MxcUri, RoomAliasId, RoomId, RoomOrAliasId, RoomVersionId, ServerName,
-    ServerSignatures, ServerSigningKeyId, SessionId, Signatures, SigningKeyAlgorithm,
-    TransactionId, UserId,
+    EventId, KeyId, KeyName, MxcUri, RoomAliasId, RoomId, RoomOrAliasId, RoomVersionId,
+    SecretRequestId, ServerName, ServerSignatures, ServerSigningKeyId, SessionId, Signatures,
+    SigningKeyAlgorithm, TransactionId, UserId,
 };
 
 #[cfg(feature = "client")]